feat: replace asynchronously loaded hacks with imported text
BREAKING CHANGE: Hacks are no longer loaded from Github with local backups, and are instead local-only. This means that if hacks are updated, Borksy must be updated as well. However, updating Borksy to pick up the latest changes in hacks is now almost fully automatic: local copies of hacks are pulled from the latest publish of the hacks repo, and all of the metadata is parsed from the hacks themselves instead of maintained manually in Borksy. A script is included (`update-hacks`) which will make fully automating hack updates possible in the future.